refactor - see message
This commit starts a bit of a refactoring of the main logic, mostly
regarding the parsing of the response.

The original style started to become a bit unwieldy, with a lot of
repetitions. The new idea is to treat the response parsers as a kind
of plugin - they are just a lambda. What they do, doesn't concern the
caller. What the caller cares about is the format of the data received.
(see below)

Another big change is the returned data format. Instead of returning the
data with native dbus types, everything is returned as dbus string, in the
form of a json object - this greatly simplifies the error handling, and
the output signatures. When constucting the json, the caller parses
the response from the modem, and expects a `map<string, string>` to
be returned. Then it iterates through it, and sends it off to the dbus-caller.